When activating an ID, 0.66 USD is transferred to the Royalty Wallet, which is equivalent to 0.33 USD, and credit will be given accordingly.
For the Bonus Travel Point, a 2.22% bonus will be credited to the Travel Point wallet. Additionally, 1.66% of the amount is credited as support, with another 1.66% (amounting to 0.83 USD) provided as promo credit. If there are any differences, 1.11 USD will be credited as promo credit to address them.

Regarding wallet transfers, there is an option to add savings, travel points, and bonus travel points directly to the bonus travel points wallet.

For transfers to savings income, you can move network income, leadership income, car & house fund, and royalty to the available balance wallet, but these transfers will follow specific given conditions.